support mode auto detect
